Move xcm and assets related runtime configurations to own files
